server/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
easysnmp==0.2.6
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mod_wsgi==5.0.2
netmiko==4.4.0
ntc_templates==6.0.0
paramiko==3.5.0
pycparser==2.22
pymongo==4.9.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
scp==0.15.0
six==1.16.0
textfsm==1.1.3
Werkzeug==3.0.4
